Offloaded Functional Arch Package
Moved functional arch ML to its own package.  It's not included in the DELS.mdzip because we haven't used it for anything in a while.

Converted the Functional blocks into activities, and fixed the dependencies. There is still some outstanding work to the done on fixing the FRN IBDs in the warehouse package. But it's not a pressing need.